modifed the HealthService and updated the test class

+++ src/main/java/com/acoustic/repository/HealthRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HealthRepository extends JpaRepository<Health, Long> {
+public interface HealthRepository extends JpaRepository<Health, Integer> {
